Implement per-file expiry in Puppet::Resource::Type
Introduce the ability to associate filenames with code added to a Type,
as well as an expire_file method to remove all code associated with a
given filename.

This feature is intended to facilitate implementation of partial AST
expiration based on an external source of information regarding changed
files.

Some of the added tests specifically match code and docs in different
orders because of changed hash behaviour between Ruby 1.8 and 1.9.