Implement automatic reloading of code using inotify
This commit introduces a new optional feature to expire (and,
optionally, reload) manifest code at runtime, based on inotify events.
The change is a noop by default, since it introduces a new
'inotify_reload' configuration parameter which defaults to false. When
switched on, this will expire code from changed files, both in the
manifests directory and from the autoloader.

This patch will currently not be fully functional with the future
parser, because the manifest directory is not recursively monitored so
some types of change will be missed. This is planned to be fixed in a
future revision.